requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
google-generativeai
pydantic
